T1759: Merging interface.py into ifconfig.py
